WIP 15:22 mar8

+++ src/com/github/group/BroadcastMessage.java
@@ -23,7 +23,7 @@ public class BroadcastMessage extends Message {
     public String username;
     public String id;
     public String ip;
-    public long port;
+    public int port;
 
     private JSONObject msg;
 
@@ -72,7 +72,7 @@ public class BroadcastMessage extends Message {
             username = (String) msg.get("Username");
             id = (String) msg.get("ID");
             ip = (String) msg.get("IP");
-            port = (long) msg.get("Port");
+            port = (int) msg.get("Port");
         } catch (ParseException e) {
             log.printLogMessage(Log.ERROR, CLASS_ID, "Received invalid BroadcastMessage");
         }
